filter like posts, users, follow feat

+++ app/controllers/likes_controller.rb
@@ -3,36 +3,36 @@ class LikesController < ApplicationController
 
   # POST /likes or /likes.json
   def create
+    unless params[:post_id].present?
+      return render json: { message: 'post_id is nil' }, status: :bad_request
+    end
 
-    l = Like.find_by like_params
+    l = Like.where(user_id: current_user.id, post_id: params[:post_id]).first
+    post = Post.find(params[:post_id])
+
     if l.present?
-      post = Post.find(like_params[:post_id])
       post.likes_count = post.likes_count - 1
       ActionCable.server.broadcast('PostsChannel', { post: post })
 
       l.destroy
       render json: {
+        likes_count: post.likes_count,
         message: 'unlike success',
-        status: 0
       }
       return
     end
 
-    like = Like.new(like_params)
+    like = Like.new({ user_id: current_user.id, post_id: params[:post_id] })
 
     if like.save
       render json: {
+        likes_count: post.likes_count.nil? ? 1 : post.likes_count + 1,
         message: 'like success',
-        status: 1
       }
     else
       render status: :unprocessable_entity
     end
   end
 
-  private
-
-  def like_params
-    params.require(:like).permit(:user_id, :post_id)
-  end
 end
+
